- refactored the mobile code for taobao and tmall
