+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jarviscrawlercore",
-  "version": "0.3.78",
+  "version": "0.3.79",
   "description": "jarvis crawler core",
   "main": "index.js",
   "scripts": {
